requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pinecone-client
# Latest llama_index with Gemini support
git+https://github.com/jerryjliu/llama_index.git@main

